xóa, đọc, ghi, jump bằng bootloader OK

+++ Boot/Bootloader/src/bl_flash.h
@@ -16,7 +16,7 @@ BlStatus_t BlFlash_Erase(uint32_t addr, uint32_t numberOfSectors);
 /* Write data */
 BlStatus_t BlFlash_Write(uint32_t addr, uint8_t *data, uint32_t len);
 
-BlStatus_t BlFlash_Read(uint32_t addr, uint32_t *data, uint32_t len);
+BlStatus_t BlFlash_Read(uint32_t addr, uint8_t *data, uint32_t len);
 
 /* Validate application */
 bool BlFlash_IsAppValid(void);
